requirements.txt
python-hcl2>=4.3.0
pyyaml>=6.0.1
click>=8.1.0
rich>=13.0.0
jinja2>=3.1.0
pytest>=7.4.0
pytest-cov>=4.1.0
